try to use robot localization
